+++ src-webviews/src/pages/chat/interfaces/ICommand.ts
@@ -1,4 +1,0 @@
-export default interface ICommand {
-    name: string;
-    description: string;
-}