+++ src-webviews/src/pages/components.ts
@@ -7,8 +7,6 @@ import Inventory from './inventory/Inventory.vue';
 import Job from './job/Job.vue';
 import MultiPageExample from './multiPageExample/MultiPageExample.vue';
 import Storage from './storage/Storage.vue';
-import Chat from './chat/Chat.vue';
-import Hud from './hud/Hud.vue';
 import Audio from './audio/Audio.vue';
 import Actions from './actions/Actions.vue';
 import AthenaLogo from './athenaLogo/AthenaLogo.vue';
@@ -20,7 +18,6 @@ export const CORE_IMPORTS = {
     Actions: shallowRef(Actions),
     AthenaLogo: shallowRef(AthenaLogo),
     Audio: shallowRef(Audio),
-    Chat: shallowRef(Chat),
     Designs: shallowRef(Designs),
     Icons: shallowRef(Icons),
     InputBox: shallowRef(InputBox),