+++ src/core/plugins/core-chat/server/src/chat.ts
@@ -1,7 +1,7 @@
 import * as alt from 'alt-server';
 
 import { Athena } from '@AthenaServer/api/athena';
-import { CHAT_CONFIG } from './config';
+import { CHAT_CONFIG } from '../../shared/config';
 
 function handleMessage(player: alt.Player, msg: string) {
     const data = Athena.document.character.get(player);
@@ -9,7 +9,7 @@ function handleMessage(player: alt.Player, msg: string) {
         return;
     }
 
-    const closestPlayers = Athena.get.players.inRange(player.pos, CHAT_CONFIG.MAX_RANGE);
+    const closestPlayers = Athena.get.players.inRange(player.pos, CHAT_CONFIG.settings.range);
     Athena.systems.messenger.players.send(closestPlayers, `${data.name}: ${msg}`);
 }