+++ src/core/plugins/core-chat/server/src/config.ts
@@ -1,4 +1,0 @@
-export const CHAT_CONFIG = {
-    
-    MAX_RANGE: 15,
-};